+++ src/Cli/Microsoft.DotNet.Cli.Definitions/Commands/New/InstantiateOptions.cs
@@ -2,9 +2,8 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.CommandLine;
-using Microsoft.TemplateEngine.Cli.Commands;
 
-namespace Microsoft.TemplateEngine.Cli;
+namespace Microsoft.DotNet.Cli.Commands.New;
 
 public sealed class InstantiateOptions
 {
