+++ src/Cli/Microsoft.DotNet.Cli.Definitions/Commands/New/TabularOutputSettingsColumnNames.cs
@@ -1,7 +1,7 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-namespace Microsoft.TemplateEngine.Cli.Commands;
+namespace Microsoft.DotNet.Cli.Commands.New;
 
 public static class TabularOutputSettingsColumnNames
 {
