+++ src/Cli/Microsoft.TemplateEngine.Cli/Commands/alias/AliasShowCommand.cs
@@ -2,13 +2,14 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.CommandLine;
+using Microsoft.DotNet.Cli.Commands.New;
 using Microsoft.TemplateEngine.Abstractions;
 using Microsoft.TemplateEngine.Edge.Settings;
 
 namespace Microsoft.TemplateEngine.Cli.Commands
 {
-    internal sealed class AliasShowCommand(Func<ParseResult, ITemplateEngineHost> hostBuilder, AliasShowCommandDefinition definition)
-        : BaseCommand<AliasShowCommandArgs, AliasShowCommandDefinition>(hostBuilder, definition)
+    internal sealed class AliasShowCommand(Func<ParseResult, ITemplateEngineHost> hostBuilder, NewAliasShowCommandDefinition definition)
+        : BaseCommand<AliasShowCommandArgs, NewAliasShowCommandDefinition>(hostBuilder, definition)
     {
         protected override Task<NewCommandStatus> ExecuteAsync(
             AliasShowCommandArgs args,
