+++ src/Cli/Microsoft.TemplateEngine.Cli/Commands/details/DetailsCommand.cs
@@ -2,14 +2,15 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.CommandLine;
+using Microsoft.DotNet.Cli.Commands.New;
 using Microsoft.TemplateEngine.Abstractions;
 using Microsoft.TemplateEngine.Cli.NuGet;
 using Microsoft.TemplateEngine.Edge.Settings;
 
 namespace Microsoft.TemplateEngine.Cli.Commands
 {
-    internal sealed class DetailsCommand(Func<ParseResult, ITemplateEngineHost> hostBuilder, DetailsCommandDefinition definition)
-        : BaseCommand<DetailsCommandArgs, DetailsCommandDefinition>(hostBuilder, definition)
+    internal sealed class DetailsCommand(Func<ParseResult, ITemplateEngineHost> hostBuilder, NewDetailsCommandDefinition definition)
+        : BaseCommand<DetailsCommandArgs, NewDetailsCommandDefinition>(hostBuilder, definition)
     {
         protected override async Task<NewCommandStatus> ExecuteAsync(
             DetailsCommandArgs args,
