+++ src/Cli/Microsoft.TemplateEngine.Cli/Commands/list/ListCommandArgs.cs
@@ -2,10 +2,11 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.CommandLine;
+using Microsoft.DotNet.Cli.Commands.New;
 
 namespace Microsoft.TemplateEngine.Cli.Commands
 {
-    internal sealed class ListCommandArgs : BaseFilterableArgs<ListCommandDefinition>, ITabularOutputArgs
+    internal sealed class ListCommandArgs : BaseFilterableArgs<NewListCommandDefinition>, ITabularOutputArgs
     {
         internal ListCommandArgs(BaseListCommand command, ParseResult parseResult)
             : base(parseResult)
