+++ src/Cli/Microsoft.TemplateEngine.Cli/Commands/search/SearchCommandArgs.cs
@@ -2,10 +2,11 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.CommandLine;
+using Microsoft.DotNet.Cli.Commands.New;
 
 namespace Microsoft.TemplateEngine.Cli.Commands
 {
-    internal sealed class SearchCommandArgs : BaseFilterableArgs<SearchCommandDefinition>, ITabularOutputArgs
+    internal sealed class SearchCommandArgs : BaseFilterableArgs<NewSearchCommandDefinition>, ITabularOutputArgs
     {
         internal SearchCommandArgs(BaseSearchCommand command, ParseResult parseResult)
             : base(parseResult)
