+++ src/Cli/Microsoft.TemplateEngine.Cli/Commands/uninstall/LegacyUninstallCommand.cs
@@ -2,12 +2,13 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.CommandLine;
+using Microsoft.DotNet.Cli.Commands.New;
 using Microsoft.TemplateEngine.Abstractions;
 using Microsoft.TemplateEngine.Edge.Settings;
 
 namespace Microsoft.TemplateEngine.Cli.Commands
 {
-    internal sealed class LegacyUninstallCommand(Func<ParseResult, ITemplateEngineHost> hostBuilder, UninstallCommandDefinition definition)
+    internal sealed class LegacyUninstallCommand(Func<ParseResult, ITemplateEngineHost> hostBuilder, NewUninstallCommandDefinition definition)
         : BaseUninstallCommand(hostBuilder, definition)
     {
         protected override Task<NewCommandStatus> ExecuteAsync(UninstallCommandArgs args, IEngineEnvironmentSettings environmentSettings, TemplatePackageManager templatePackageManager, ParseResult parseResult, CancellationToken cancellationToken)
