+++ src/Cli/Microsoft.TemplateEngine.Cli/Commands/update/BaseUpdateCommand.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.CommandLine;
+using Microsoft.DotNet.Cli.Commands.New;
 using Microsoft.TemplateEngine.Abstractions;
 using Microsoft.TemplateEngine.Edge.Settings;
 
@@ -9,14 +10,14 @@ namespace Microsoft.TemplateEngine.Cli.Commands
 {
     internal interface IUpdateCommand
     {
-        UpdateCommandDefinitionBase Definition { get; }
+        NewUpdateCommandDefinitionBase Definition { get; }
     }
 
     internal abstract class BaseUpdateCommand<TDefinition>(Func<ParseResult, ITemplateEngineHost> hostBuilder, TDefinition definition)
         : BaseCommand<UpdateCommandArgs, TDefinition>(hostBuilder, definition), IUpdateCommand
-        where TDefinition : UpdateCommandDefinitionBase
+        where TDefinition : NewUpdateCommandDefinitionBase
     {
-        UpdateCommandDefinitionBase IUpdateCommand.Definition => Definition;
+        NewUpdateCommandDefinitionBase IUpdateCommand.Definition => Definition;
 
         protected override Task<NewCommandStatus> ExecuteAsync(
             UpdateCommandArgs args,
