+++ src/Cli/Microsoft.TemplateEngine.Cli/Commands/update/UpdateCommand.cs
@@ -2,13 +2,14 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.CommandLine;
+using Microsoft.DotNet.Cli.Commands.New;
 using Microsoft.TemplateEngine.Abstractions;
 using Microsoft.TemplateEngine.Edge.Settings;
 
 namespace Microsoft.TemplateEngine.Cli.Commands
 {
-    internal sealed class UpdateCommand(Func<ParseResult, ITemplateEngineHost> hostBuilder, UpdateCommandDefinition definition)
-        : BaseUpdateCommand<UpdateCommandDefinition>(hostBuilder, definition)
+    internal sealed class UpdateCommand(Func<ParseResult, ITemplateEngineHost> hostBuilder, NewUpdateCommandDefinition definition)
+        : BaseUpdateCommand<NewUpdateCommandDefinition>(hostBuilder, definition)
     {
         protected override async Task<NewCommandStatus> ExecuteAsync(
             UpdateCommandArgs args,
